requirements.txt
opencv-python
numpy
pandas
Pillow